+++ App.js
@@ -18,7 +18,7 @@ import EventsPage from './src/components/EventsPage';
 import Homepage from './src/components/Homepage';
 import MembershipPage from './src/components/MembershipPage';
 import EventDetailPage from './src/components/EventDetailsPage';
-import SignUp from './src/components/SignUp';
+import KeepInTouch from './src/components/KeepInTouch';
 import PhoneCall from './src/components/PhoneCall';
 import PhoneSMS from './src/components/PhoneSMS';
 
@@ -89,7 +89,7 @@ export default function App() {
         <Stack.Screen name='Workshops' component={EventsPage} />
         <Stack.Screen name='EventDetailPage' component={EventDetailPage} />
         <Stack.Screen name='Community' component={MembershipPage} />
-        <Stack.Screen name='SignUp' component={SignUp} />
+        <Stack.Screen name='KeepInTouch' component={KeepInTouch} />
       </Stack.Navigator>
     </NavigationContainer>
   ) : (
